refactor: auth.ts 파일 이동
